+++ bin/main.dart
@@ -7,7 +7,6 @@ main() {
     String currentUrl = document.baseUri;
     List<String> urlParts = path.split(currentUrl);
     sg2o.Styles styles = new sg2o.Styles();
-    sg2o.MyStorage storage = new sg2o.MyStorage();
     sg2o.SideBar sidebar = new sg2o.SideBar();
 
     sg2o.activateSettings();
@@ -29,6 +28,7 @@ main() {
         }
 
         sidebar.addFilter();
+        sidebar.moveAdsToTheEnd();
     }
 
     if (sidebar.isSidebarPresent()) {