+++ lib/constants.dart
@@ -19,10 +19,21 @@ const String classFeaturedContainer = 'featured__container';
 const String classFeaturedOuterWrap = 'featured__outer-wrap--home';
 const String classWidgetContainers = 'widget-container';
 const String classSectionHeading = 'page__heading';
+const String classSectionDescription = 'page__description';
+const String classPageOuterWrap = 'page__outer-wrap';
 const String classPoints = 'nav__points';
 const String classLeftNavigation = 'nav__left-container';
 const String classNavigationButtonContainer = 'nav__button-container';
+const String classNavigationRelativeDropDown = 'nav__relative-dropdown';
+const String classNavigationAbsoluteDropDown = 'nav__absolute-dropdown';
 const String classNavigationButton = 'nav__button';
+const String classNavigationButtonIsDropDown = 'nav__button--is-dropdown';
+const String classNavigationButtonIsDropDownArrow = 'nav__button--is-dropdown-arrow';
+const String classNavigationRow = 'nav__row';
+const String classNavigationRowSummary = 'nav__row__summary';
+const String classNavigationRowSummaryName = 'nav__row__summary__name';
+const String classNavigationRowSummaryDescription = 'nav__row__summary__description';
+const String classFAAngleDown = 'fa-angle-down';
 const String classSidebar = 'sidebar';
 const String classSidebarHeading = 'sidebar__heading';
 
@@ -41,13 +52,17 @@ const String classGAContributorNeg = 'giveaway__column--contributor-level--negat
 const String classGAGroupIcon = 'giveaway__column--group';
 const String classGAEntered = 'is-faded';
 const String classGAImageOuterBorder = 'global__image-outer-wrap';
+const String classGAImageOuterBorderGameSmall = 'global__image-outer-wrap--game-small';
+const String classGAImageInnerWrap = 'global__image-inner-wrap';
 const String classGAWhiteListed = 'giveaway__column--whitelist';
 const String classGAHide = 'giveaway__hide';
+const String classGAIcon = 'giveaway__icon';
 const String classFontAwesome = 'fa';
 const String classFAFullHeart = 'fa-heart';
 const String classFAEmptyHeart = 'fa-heart-o';
 const String classEyeSlash = 'fa-eye-slash';
 const String classFrown = 'fa-frown-o';
+const String classFASteam = 'fa-steam';
 
 const String classWidgetMarginTop = 'widget-container--margin-top';
 
@@ -56,9 +71,18 @@ const String classPaginationNavigation = 'pagination__navigation';
 const String classPageHeadingBreadCrumbs = 'page__heading__breadcrumbs';
 
 const String classTableColumnWidthFill = 'table__column--width-fill';
+const String classTableColumnWidthSmall = 'table__column--width-small';
 const String classTableHeading = 'table__heading';
+const String classTableRemoveDefault = 'table__remove-default';
+const String classTableColumnSecondaryLink = 'table__column__secondary-link';
+const String classTableColumnHeading = 'table__column__heading';
+const String classTableRows = 'table__rows';
+const String classTableRowOuterWrap = 'table__row-outer-wrap';
+const String classTableRowInnerWrap = 'table__row-inner-wrap';
 
 const String classWishListName = 'table__column__heading';
+
+const String classTextCenter = 'text-center';
 
 const String classFormRow = 'form__row';
 const String classFormHeading = 'form__heading';
@@ -77,6 +101,8 @@ const String classFASpin = 'fa-spin';
 
 const String classIsSelected = 'is-selected';
 const String classIsDisabled = 'is-disabled';
+const String classIsHidden = 'is-hidden';
+const String classIsClickable = 'is-clickable';
 
 /// elements defined by sg
 
@@ -97,6 +123,7 @@ const String classGridViewGAImage = 'sg2o-game-image';
 const String classOneClickBlackList = 'sg2o-blacklist-ga';
 const String classCustomWishList = 'sg2o-custom-wishlist';
 const String classGiveAwayBlackList = 'sg2o-giveaway-blacklist';
+const String classSteamLink = 'sg2o-steam-link';
 const String classHighLightCustomWishList = 'sg2o-custom-wishlist-red';
 const String classTooltip = 'sg2o-tooltip';
 const String classProfileTooltip = 'sg2o-profile-tooltip';
@@ -144,6 +171,13 @@ const String classSettingsTabContent = 'sg2o-settings-tabs-content';
 const String classSettingsTabSelected = 'sg2o-settings-tab-selected';
 const String classSettingsColorInput = 'sg2o-settings-color-input';
 
+const String classSG2ORecentlyBlackListedLink = 'sg2o-overlay-recently-blacklisted';
+const String classSG2OAboutLink = 'sg2o-overlay-about';
+
+/// IDs
+// Add game name to this id.
+const String idBlacklistHistoryRow = 'sg2o-recently-blacklisted-entry-';
+
 /// URLs
 const String urlSteamGiftsHome = 'http://www.steamgifts.com/';
 const String urlGiveAwayPage = '/giveaways/search?page=';
@@ -163,6 +197,8 @@ const String keyAutomaticBlackList = 'automatic-blacklist';
 const String keyWishList = 'wishlist';
 const String keyCustomWishList = 'custom-wishlist';
 const String keyGiveAwayBlackList = 'giveaway-blacklist';
+
+const String keyRecentlyBlacklistedGiveaways = 'giveaway-blacklist-recently';
 
 const String oldFilterKey = 'lscache-sgpgiveawayFilters';
 
@@ -187,3 +223,5 @@ const num oneDay = 1000 * 60 * 60 * 24;
 
 const num gridViewEntryWidth = 196;
 const num gridViewEntryHeight = 81;
+
+const num zIndexLayer = 100000; //needs to be 99.999+1 because of the filter widget.