+++ lib/giveaway/giveaway.dart
@@ -17,6 +17,8 @@ class GiveAway {
     int comments;
     int contributorLevel = 0;
     int sgGameId;
+    int steamId;
+    String steamLink;
     num chanceOfWin;
     bool isContributorGA;
     bool isGroupGA;
@@ -78,10 +80,20 @@ class GiveAway {
         this.entered = gaHtml.querySelectorAll('.$classGAEntered').length > 0;
         this.isWhiteListed = gaHtml.querySelectorAll('.$classGAWhiteListed').length > 0;
         this.sgGameId = parseNumber(gaHtml.getAttribute('data-game-id'));
+        this.steamLink = getSteamLink(gaHtml);
+        this.steamId = parseNumber(this.steamLink);
 
         this.chanceOfWin = ((100 * this.copies) / (this.entries + ((entered) ? 0 : 1))).clamp(0, 100);
 
         this.borderClass = getBorderColorClass();
+    }
+
+    String getSteamLink(Element gaHtml) {
+        Element querySelector = gaHtml.querySelector('a.$classGAIcon');
+        if (querySelector == null) {
+            return '';
+        }
+        return querySelector.getAttribute('href');
     }
 
     /// Return this giveaway as small gridview compatible [Element].
@@ -173,7 +185,7 @@ class GiveAway {
         DivElement avatarContainer = new DivElement()
             ..classes.add(classFloatRight)
             ..classes.add(classGridViewAvatar)
-            ..id = 'sg2o-$creator-$giveAwayId'
+            ..id = 'sg2o-$creator-$giveAwayId-avatar'
             ..attributes.putIfAbsent('style', () => 'background-image: ${this.avatar};')
             ..onClick.listen((Event e) {
                 window.open(this.profileLink, '_blank');
@@ -184,7 +196,7 @@ class GiveAway {
         giveAwayContainer
             ..onMouseEnter.listen((Event e) {
                 if (profileTooltip == null) {
-                    profileTooltip = new ProfileTooltip('sg2o-$creator-$giveAwayId', creator, avatarContainer);
+                    profileTooltip = new ProfileTooltip('sg2o-$creator-$giveAwayId-avatar', creator, avatarContainer);
                 }
             });
         // Only load the profile info when hovering over the avatar to reduce load on SG
@@ -230,23 +242,23 @@ class GiveAway {
             ..classes.add(classFontAwesome)
             ..classes.add(classEyeSlash)
             ..classes.add(classOneClickBlackList)
-            ..classes.add(classTooltip)
-            ..append(new SpanElement()..innerHtml = '<b></b>Add ${this.name} to the blacklist.')
+            ..id = 'sg2o-$creator-$giveAwayId-blacklist'
             ..onClick.listen((Event e) {
                 addGameToBlackList();
                 gridView.removeByName(this.name);
             })
         ;
+        new SimpleTooltip('sg2o-$creator-$giveAwayId-blacklist', 'Add "${this.name}" to the blacklist.', blackListLinkContainer);
 
         // Add [DivElement] to add game directly to custom wishlist.
         DivElement customWishListContainer = new DivElement()
             ..classes.add(classFloatLeft)
             ..classes.add(classFontAwesome)
             ..classes.add(classCustomWishList)
-            ..classes.add(classTooltip)
-            ..append(new SpanElement()..innerHtml = '<b></b>Add ${this.name} to a custom wishlist')
+            ..id = 'sg2o-$creator-$giveAwayId-custom-wishlist'
             ..onClick.listen(toggleGameOnCustomWishList)
         ;
+        new SimpleTooltip('sg2o-$creator-$giveAwayId-custom-wishlist', 'Add "${this.name}" to a custom wishlist', customWishListContainer);
 
         // Add [DivElement] to add giveaway directly to the [GiveAwayBlackList].
         DivElement giveAwayBlackListContainer = new DivElement()
@@ -254,8 +266,7 @@ class GiveAway {
             ..classes.add(classFontAwesome)
             ..classes.add(classFrown)
             ..classes.add(classGiveAwayBlackList)
-            ..classes.add(classTooltip)
-            ..append(new SpanElement()..innerHtml = '<b></b>Hide this giveaway with ID <em>${this.giveAwayId}</em> until it is finished.')
+            ..id = 'sg2o-$creator-$giveAwayId-id-blacklist'
             ..onClick.listen((Event e) {
                 giveAwayBlackList.addGameToBlackList(this.giveAwayId);
                 // remove giveaway from gridview so it is not added back again
@@ -264,6 +275,20 @@ class GiveAway {
                 giveAwayContainer.remove();
             })
         ;
+        new SimpleTooltip('sg2o-$creator-$giveAwayId-id-blacklist', 'Hide this giveaway with ID "${this.giveAwayId}" until it is finished.', giveAwayBlackListContainer);
+
+        // Add [DivElement] to open the Steam store page.
+        DivElement steamLinkContainer = new DivElement()
+            ..classes.add(classFloatLeft)
+            ..classes.add(classFontAwesome)
+            ..classes.add(classSteamLink)
+            ..classes.add(classFASteam)
+            ..id = 'sg2o-$creator-$giveAwayId-steam'
+            ..onClick.listen((Event e) {
+                window.open(this.steamLink, '_blank');
+            })
+        ;
+        new SimpleTooltip('sg2o-$creator-$giveAwayId-steam', 'Open Steam page for this game.', steamLinkContainer);
 
         if (isCustomWishListGA) {
             customWishListContainer.classes.add(classFAFullHeart);
@@ -285,6 +310,7 @@ class GiveAway {
             blackListLinkContainer.classes.add(classFaded);
             customWishListContainer.classes.add(classFaded);
             giveAwayBlackListContainer.classes.add(classFaded);
+            steamLinkContainer.classes.add(classFaded);
         }
 
         informationContainer
@@ -304,6 +330,7 @@ class GiveAway {
             ..append(blackListLinkContainer)
             ..append(customWishListContainer)
             ..append(giveAwayBlackListContainer)
+            ..append(steamLinkContainer)
         ;
 
         return informationContainer;
@@ -316,11 +343,6 @@ class GiveAway {
             giveAwayContainer.remove();
             this.isBlackListed = true;
         }
-    }
-
-    /// Print this giveaway to console.
-    void print() {
-        window.console.info(toString());
     }
 
     /// Returns a [String]-representation of this giveaway.
@@ -381,7 +403,8 @@ class GiveAway {
     /// Sends a post request to directly add the game on SGs blacklist.
     void addGameToBlackList() {
         Map<String, String> formData = new Map();
-        formData['xsrf_token'] = querySelectorAll('input[name="xsrf_token"]')[0].getAttribute('value');
+        String xsrfToken = querySelectorAll('input[name="xsrf_token"]')[0].getAttribute('value');
+        formData['xsrf_token'] = xsrfToken;
         formData['game_id'] = this.sgGameId.toString();
         formData['do'] = 'hide_giveaways_by_game_id';
         HttpRequest.postFormData('/ajax.php', formData);