+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.3.0
+- Added steam link to the giveaway entries on the gridview (#90)
+- Added navigation menu to the sg2o button in the top bar
+  - SG₂O button leads to the discussion thread
+  - in the dropdown there are links to settings and a simple about page (#123)
+- Using PowerTip (jQuery) instead of custom tooltip (#109)
+- Fixed an issue with games not having a steam link (#149)
+- Fixed settings can't be deactivated (#155)
+- Move left side ads below filters (#154)
+
 ## 0.2.21
 - Fix gridview after some restyling of cg.
 - Show colored borders correctly